refactor(cli): extract drt/cli/commands/ package — Phase 1 of cli/main.py split (closes #546)
Ninth and final slice of the Tech Foundation Hardening epic (#538).

Per the issue's own guidance ("One PR per 'wave' of moves … to keep
review surface tractable"), this lands Phase 1: the package scaffold
plus the simplest, most self-contained commands. Heavier commands
(run / validate / status / test / init / serve) stay in main.py for a
Phase 2 follow-up issue — see ``Out of scope`` below.

What moves
----------

New ``drt/cli/_app.py`` holds the shared ``typer.Typer`` instance so
per-command modules can ``from drt.cli._app import app`` without
triggering a circular import against ``drt/cli/main.py``.

New ``drt/cli/commands/`` package with six modules:

- ``commands/doctor.py``       — ``drt doctor``
- ``commands/list_syncs.py``   — ``drt list``
- ``commands/config.py``       — ``drt config set/unset/show-telemetry``
- ``commands/cloud.py``        — ``drt cloud push/status`` (stubs)
- ``commands/docs.py``         — ``drt docs generate/serve`` (epic #499)
- ``commands/mcp.py``          — ``drt mcp run``

Each command module re-uses the shared ``app``; the four sub-Typer
namespaces (config / cloud / docs / mcp) define their own
``<name>_app = typer.Typer(...)`` and call ``app.add_typer(<name>_app)``
at module-import time. The ``drt.cli.commands.__init__`` side-imports
every module so registration happens when ``drt/cli/main.py`` imports
the package.

``main.py`` updates
-------------------

- Drops the inlined ``app = typer.Typer(...)`` definition in favour of
  ``from drt.cli._app import app`` so the new modules see the same
  instance.
- Adds ``from drt.cli import commands as _commands`` to trigger the
  registration side effects on import.
- Removes the migrated command bodies (~240 LOC) and replaces each
  removed block with a one-line pointer comment so future readers
  find the new home immediately.

main.py: **1706 → 1470 LOC** (-236, -14%).

Backward compat
---------------

Tests import ``from drt.cli.main import app`` (10+ call sites). The
re-export still works because ``main.py`` continues to import ``app``
from ``_app`` into its module namespace. ``_run_one``, ``_RunContext``,
``_get_destination``, ``_get_source``, ``_configure_json_logging``,
``_JsonFormatter``, ``_test_display_name`` — every other test-imported
symbol stays put for Phase 2.

Out of scope (Phase 2 — separate follow-up issue)
-------------------------------------------------

Heavier commands stay in main.py for a future PR:

- ``init`` (~290 LOC + ``_init_from_dbt`` / ``_list_templates`` /
  ``_init_from_template`` helpers)
- ``sources`` / ``destinations`` (~150 LOC + the connector printer
  helpers — natural pair to move with the #543 detail builder)
- ``clean`` (~50 LOC, depends on ``_get_destination``)
- ``run`` + ``_RunContext`` + ``_run_one`` + ``_print_watermark_summary``
  (~450 LOC, heaviest, threading + signal handling)
- ``validate`` (~200 LOC, secrets scan + connection check)
- ``status`` (~120 LOC, StateManager interplay)
- ``test_syncs`` (~130 LOC, sync test framework)
- ``serve`` (small, MCP-adjacent)

Each is mechanical now that the pattern exists — copy to
``commands/<name>.py``, change ``@app.command()`` source to
``from drt.cli._app import app``, add to the ``__init__.py`` side-
imports. The split was sequenced this way so the high-risk commands
(run / validate / status / test) move with their own dedicated review
attention rather than being buried in a 2000-line refactor diff.

Local verification
------------------

  pytest tests/                  → 1185 passed, 1 skipped, 7 deselected
  ruff check drt tests           → All checks passed
  mypy drt                       → Success: no issues in 104 files

Refs: #538 (epic — this closes the last Wave 2 issue), #546 (this
issue), #565 (TBD follow-up for Phase 2 commands).
