+++ drt/cli/commands/doctor.py
@@ -1,0 +1,13 @@
+"""`drt doctor` — environment check command."""
+
+from __future__ import annotations
+
+from drt.cli._app import app
+
+
+@app.command()
+def doctor() -> None:
+    """Check environment and report potential issues."""
+    from drt.cli.doctor import run_doctor
+
+    run_doctor()
